refactor: extract run-level action handlers from ProjectPage into useRunActions hook
Move 12 action handlers (approve, generate, restart, render, stop,
resume, delete, go-back, title-save + visual plan variants) and their
16 associated state variables into a dedicated useRunActions hook.
ProjectPage reduced from 536 to 252 lines.

Closes #370
